address convoy teardown review feedback
